requirements.txt
streamlit
sqlalchemy
google-generativeai
python-dotenv